+++ tests/textindex/MSexceptions.sql
@@ -1,0 +1,1 @@
+CREATE EXCEPTION e_no_folder 'Folder not found';